// src/app/models/menu.model.ts
export class MenuModel {
  CodigoSolidoInicial: string;
  CodigoLiquidoInicial: string;
  ProductoSolidoInicial: string;
  FrecuenciaSolidoUtilizadoInicial: number;
  frecuenciaSolidoInicialInicial: number;
  frecuenciaLiquidaInicialInicial: number;
  montoLiquidaInicial: number;
  MontoSolildoUtilizadoInicial: number;
  montoSolildoInicial: number;
  ProductoLiquidoInicial: string;
  FrecuenciaLiquidaUtilizadoInicial: number;
  MontoLiquidaUtilizadoInicial: number;
  CodigoSolidoPrimaria: string;
  CodigoLiquidoPrimaria: string;
  ProductoSolidoPrimaria: string;
  FrecuenciaSolidoUtilizadoPrimaria: number;
  MontoSolildoUtilizadoPrimaria: number;
  ProductoLiquidoPrimaria: string;
  FrecuenciaLiquidaUtilizadoPrimaria: number;
  MontoLiquidaUtilizadoPrimaria: number;
  frecuenciaSolidoPrimariaInicial: number;
  montoSolildoPrimaria: number;
  frecuenciaLiquidaPrimariaInicial: number;
  montoLiquidaPrimaria: number;
  CodigoSolidoSegundaria: string;
  CodigoLiquidoSegundaria: string;
  ProductoSolidoSegundaria: string;
  FrecuenciaSolidoUtilizadoSegundaria: number;
  MontoSolildoUtilizadaSegundaria: number;
  ProductoLiquidoSegundaria: string;
  FrecuenciaLiquidaUtilizadoSegundaria: number;
  frecuenciaSolidoSegundariaInicial: number;
  montoSolildoSegundaria: number;
  frecuenciaLiquidaSegundariaInicial: number;
  montoLiquidaSegundaria: number;
  MontoLiquidaUtilizadoSegundaria: number;
  ObservacionJefeUnace: string;
  ObservacionEb: string;
}